backend/requirements.txt
fastapi
uvicorn
pdf2image
opencv-python
easyocr
paddleocr
python-multipart
supabase
python-dotenv
